# noisy-ninjas/backend/Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Noisy Ninjas routes documentation: powered by Swagger. JSON file built of tutorial: https://medium.com/bb-tutorials-and-thoughts/how-to-add-swagger-to-nodejs-rest-api-7a542cfdc5e1",
    "version": "1.0.0",
    "title": "Noisy Ninjas",
    "contact": {
      "email": "noisyninjasthegame@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http"],
  "host": "https://noisy-ninjas.nn.r.appspot.com/",
  "basePath": "/api",
  "paths": {
...
    "/api/users/{displayName}": {
      "get": {
        "summary": "Returns a user",

        "produces": ["application/json"],
        "parameters": [
          {
            "name": "displayName",
            "in": "path",
            "description": "user requested",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "user found",
            "schema": {
              "type": "object",
              "schema": {
                "type": "array",
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "description": "access denied"
          },
          "500": {
            "description": "unexpected conditon"
          }
        }
      }
    },
    "/api/rankings": {
      "get": {
        "summary": "Get rankings of users",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "page for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "path",
            "description": "limit for pagination",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "200": {
            "description": "rankings found",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/rankings/{displayName}": {
      "get": {
        "summary": "Get rankings of users",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "page for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "path",
            "description": "limit for pagination",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "404": {
            "description": "displayName not found"
          },
          "200": {
            "description": "rankings found",
            "schema": {
              "type": "number"
            }
          }
        }
      }
    },
    "/api/users/{displayName}/": {
      "delete": {
        "summary": "delete user by displayName",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "displayName",
            "in": "path",
            "description": "name of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "200": {
            "description": "user deleted"
          }
        }
      }
    },
    "/api/users/{displayName}/password": {
      "delete": {
        "summary": "update password of user",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "displayName",
            "in": "path",
            "description": "name of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "404": {
            "description": "User with displayName not found"
          },
          "200": {
            "description": "user found",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/users/:displayName/stats": {
      "get": {
        "summary": "get stats of user",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "displayName",
            "in": "path",
            "description": "name of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "404": {
            "description": "User with displayName not found"
          },
          "200": {
            "description": "user found",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/users/:displayName/username": {
      "patch": {
        "summary": "update username of user",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "displayName",
            "in": "path",
            "description": "name of user",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "body",
            "description": "new name for user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "403": {
            "description": "New name already exists"
          },
          "200": {
            "description": "displayName updated",
            "type": "string",
            "example": "bobby"
          }
        }
      }
    },
...
    "/signup": {
      "post": {
        "summary": "creates a user",
        "produces": ["application/json"],

        "body": [
          {
            "name": "username",
            "in": "body",
            "description": "new name for user",
            "required": true,
            "type": "string",
            "example": "bobby"
          },
          {
            "name": "password",
            "in": "body",
            "description": "new name for user",
            "required": true,
            "type": "string",
            "example": "Password1_123"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "422": {
            "description": "password or displayName missing"
          },
          "200": {
            "description": "User added",
            "type": "string",
            "example": "bobby"
          }
        }
      }
    },
    "/signin": {
      "post": {
        "summary": "signs the user in",
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "username",
            "in": "body",
            "description": "new name for user",
            "required": true,
            "type": "string",
            "example": "bobby"
          },
          {
            "name": "password",
            "in": "body",
            "description": "new name for user",
            "required": true,
            "type": "string",
            "example": "Password1_123"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "401": {
            "description": "access denied"
          },
          "422": {
            "description": "password or displayName missing"
          },
          "200": {
            "description": "User added",
            "type": "string",
            "example": "bobby"
          }
        }
      }
    },
...
    "/match": {
      "get": {
        "summary": "Gets a match from the database",
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "matchID",
            "in": "body",
            "description": "ID of match",
            "required": true,
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "200": {
            "description": "match found",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Match"
            }
          }
        }
      }
    },
...
    "/match/": {
      "delete": {
        "summary": "Deletes a match from database",
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "matchID",
            "in": "body",
            "description": "ID of match",
            "required": true,
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "500": {
            "description": "unexpected condition"
          },
          "200": {
            "description": "match has been deleted",
            "schema": {
              "type": "string",
              "example": "Match has been deleted"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "googleID": {
          "type": "string",
          "example": "10001000010"
        },
        "displayName": {
          "type": "string",
          "example": "Bobby"
        },
        "imageUrl": {
          "type": "string",
          "example": "google.com/catimg"
        },
        "points": {
          "type": "number",
          "example": 5
        },
        "gamesPlayed": {
          "type": "number",
          "example": 5
        },
        "gamesWon": {
          "type": "number",
          "example": 4
        },
        "beltRank": {
          "type": "string",
          "example": "#FFFFFF"
        },
        "matchId": {
          "type": "string",
          "example": "507f1f77bcf86cd799439011"
        },
        "hash": {
          "type": "string",
          "example": "sd][nqwdm"
        }
      }
    },
    "Match": {
      "type": "object",
      "properties": {
        "matchMap": {
          "schema": {
            "type": "array",
            "$ref": "#/definitions/Map"
          }
        },
        "matchNinjas": {
          "type": "array",
          "example": [
            { "name": "Andy", "skin": "1ninja" },
            { "name": "Andy5", "skin": "2ninja" },
            { "name": "b", "skin": "3ninja" },
            { "name": "a", "skin": "4ninja" }
          ]
        },
        "matchMonsters": {
          "type": "array",
          "example": [{ "name": "Calvin", "skin": "1monster" }]
        }
      }
    },
    "Map": {
      "type": "array",
      "example": ["#3D8B00", []]
    }
  }
}